src/requirements.txt
Pillow==9.5.0
geopy==2.3.0
python-telegram-bot==20.3
Requests==2.31.0
selenium==4.9.1
urllib3==1.26.6
PyVirtualDisplay==3.0
